backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-dotenv==1.0.0

# HTTP Client
httpx==0.28.1
requests==2.31.0

# Discord Integration
discord.py==2.3.2

# AI (OpenAI)
openai==1.57.4

# Background Jobs
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Solana Integration
solana==0.36.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring (Optional)
sentry-sdk==1.38.0
prometheus-client==0.19.0
